partial networking stack overhaul  - a lot of coding style changes (comments, indentation, etc.)  - convert several ints to unsigned ints or size_t values  - streamline many of the IPC-related macros (they no longer dereference the call structure by themselves)  - get rid of netif_interface.h (containing only aliases for remote functions and not serving any purpose)  - rename netif_local.{c|h} to netif_skel.{c|h} (it is really just a skeleton)  - drop the "_remote" and "_standalone" suffixes from most of the netif_ functions (they do not serve any purpose anymore)  - implement netif_client_connection() as a common framework function for all netif modules    - update the lo module accordingly  - ip module now reports the default gateway to the user whenever it is being set

+++ uspace/srv/net/net/net.h
@@ -134,9 +134,9 @@ typedef struct {
 
 extern int add_configuration(measured_strings_t *, const uint8_t *,
     const uint8_t *);
-extern int net_module_message(ipc_callid_t, ipc_call_t *, ipc_call_t *, int *);
+extern int net_module_message(ipc_callid_t, ipc_call_t *, ipc_call_t *, size_t *);
 extern int net_initialize_build(async_client_conn_t);
-extern int net_message(ipc_callid_t, ipc_call_t *, ipc_call_t *, int *);
+extern int net_message(ipc_callid_t, ipc_call_t *, ipc_call_t *, size_t *);
 
 #endif
